fix(ngrx-1): reducers missing from config
Due to a recent refactor of upgrading to Angular v16, the reducers weren't
copied over to the new `app.config.ts` file for the `ngrx-1` module.

Also fixed a minor spelling mistake in the related README.md file.
